feat: Implement role-based access control for reports and admin pages
