+++ src/app/[lang]/account/page.tsx
@@ -2,7 +2,8 @@ import { getDictionary } from '@/get-dictionary';
 import AccountClient from './account-client';
 import { type Locale } from '@/i18n-config';
 import { type Metadata } from 'next';
-
+import { auth } from '@/server/auth';
+import { redirect } from 'next/navigation';
 export async function generateMetadata({
     params: { lang }
 }: {
@@ -20,6 +21,10 @@ export default async function AccountPage({
 }: {
     params: { lang: Locale };
 }) {
+    const session = await auth();
+    if (!session) {
+        return redirect(`/${lang}/login`);
+    }
     const dict = await getDictionary(lang);
 
     return <AccountClient params={{ lang }} dict={dict} />;
